+++ apps/cli/src/utils.test.ts
@@ -1,8 +1,8 @@
-import { trustedSourcesValidatorMiddleware } from './utils';
-import { getTrustedSources } from './storage';
+import { trustedSourcesValidatorMiddleware } from './commands/TrustedSources';
+import { getCustomTrustedSources } from './storage';
 
 jest.mock('./storage', () => ({
-  getTrustedSources: jest.fn(() => ['https://expo.io/**']),
+  getTrustedSources: jest.fn(() => ['https://expo.dev/**']),
 }));
 
 describe('trustedSourcesValidatorMiddleware', () => {
@@ -17,27 +17,27 @@ describe('trustedSourcesValidatorMiddleware', () => {
   });
 
   it('should not throw if there are no trusted sources', async () => {
-    (getTrustedSources as jest.Mock).mockReturnValue(undefined);
+    (getCustomTrustedSources as jest.Mock).mockReturnValue(undefined);
     const fn = jest.fn();
-    await trustedSourcesValidatorMiddleware(fn)('https://expo.io/test');
-    expect(fn).toHaveBeenCalledWith('https://expo.io/test');
-    expect(getTrustedSources).toHaveBeenCalled();
+    await trustedSourcesValidatorMiddleware(fn)('https://expo.dev/test');
+    expect(fn).toHaveBeenCalledWith('https://expo.dev/test');
+    expect(getCustomTrustedSources).toHaveBeenCalled();
   });
 
   it('should not throw an error if the URL is from a trusted source', async () => {
     const fn = jest.fn();
-    await trustedSourcesValidatorMiddleware(fn)('https://expo.io/test');
-    expect(fn).toHaveBeenCalledWith('https://expo.io/test');
+    await trustedSourcesValidatorMiddleware(fn)('https://expo.dev/test');
+    expect(fn).toHaveBeenCalledWith('https://expo.dev/test');
   });
 
   it('should throw an error if the URL is from an untrusted source', async () => {
     const fn = jest.fn();
 
     try {
-      await trustedSourcesValidatorMiddleware(fn)('https://expo.io/test');
+      await trustedSourcesValidatorMiddleware(fn)('https://expo.dev/test');
     } catch (error: any) {
       expect(error).toBeInstanceOf(Error);
-      expect(error.message).toBe('This URL is from an untrusted source: https://expo.io/test');
+      expect(error.message).toBe('This URL is from an untrusted source: https://expo.dev/test');
       expect(fn).not.toHaveBeenCalled();
     }
   });