+++ apps/cli/src/utils.ts
@@ -2,8 +2,6 @@ import { InternalError } from 'common-types';
 import { Platform } from 'common-types/build/cli-commands';
 import { Env } from 'eas-shared';
 import util from 'util';
-import picomatch from 'picomatch';
-import { getTrustedSources } from './storage';
 
 export function returnLoggerMiddleware(fn: (...args: any[]) => any | Promise<any>) {
   return async function (...args: any[]) {
@@ -60,28 +58,6 @@ export function returnLoggerMiddleware(fn: (...args: any[]) => any | Promise<any
     }
   };
 }
-
-export const trustedSourcesValidatorMiddleware = (fn: (...args: any[]) => any | Promise<any>) => {
-  return async function (...args: any[]) {
-    const urls = args.filter((arg) => typeof arg === 'string' && arg.startsWith('http'));
-    if (!urls) {
-      return fn(...args);
-    }
-
-    const trustedSources = getTrustedSources();
-    if (!trustedSources) {
-      return fn(...args);
-    }
-
-    for (const url of urls) {
-      if (!trustedSources.some((source) => picomatch(source)(url))) {
-        throw new Error(`This URL is from an untrusted source: ${url}`);
-      }
-    }
-
-    return fn(...args);
-  };
-};
 
 export const getPlatformFromURI = (uri: string) => {
   if (uri.endsWith('.apk')) {