+++ apps/menu-bar/src/commands/getTrustesSourcesAsync.ts
@@ -1,7 +1,7 @@
 import MenuBarModule from '../modules/MenuBarModule';
 
 export const getTrustedSourcesAsync = async () => {
-  const trustedSources = await MenuBarModule.runCli('get-trusted-sources', [], console.log);
+  const trustedSources = await MenuBarModule.runCli('get-custom-trusted-sources', [], console.log);
 
   if (typeof trustedSources === 'string') {
     try {