+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### 🎉 New features
 
-- Add support for configuring trusted sources for opening URLs via the CLI and menu-bar app. ([#270](https://github.com/expo/orbit/pull/270) by [@altaywtf](https://github.com/altaywtf))
+- Add support for configuring trusted sources for opening URLs via the CLI and menu-bar app. ([#270](https://github.com/expo/orbit/pull/270), [#275](https://github.com/expo/orbit/pull/275) by [@altaywtf](https://github.com/altaywtf))
 
 ### 🐛 Bug fixes